新增 - 【招聘】获取角色列表 - 【招聘】获取招聘官网推广渠道列表 - 【招聘】删除招聘官网推广渠道 - 【招聘】新建招聘官网推广渠道 - 【招聘】更新招聘官网推广渠道 - 【招聘】获取地址列表 - 【招聘】获取招聘官网投递任务结果 - 【招聘】根据简历附件创建招聘官网投递任务 - 【招聘】查询地点列表 - 【招聘】恢复投递 - 【飞书人事】为用户授权角色 - 【飞书人事】更新用户被授权的数据范围 - 【消息与群组】添加跟随气泡
更新
- 获取角色详情 响应体
- 操作候选人入职 响应体
- 批量查询员工信息 响应体
- 搜索员工信息 响应体

增加权限要求描述。
变更时间戳类型 int=>long

+++ src/Ccm/GetDriveV1FilesByFileTokenCommentsByCommentIdRepliesResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-12
 // ************************************************************************
 // <copyright file="GetDriveV1FilesByFileTokenCommentsByCommentIdRepliesResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -55,7 +55,7 @@ public record GetDriveV1FilesByFileTokenCommentsByCommentIdRepliesResponseDto
         /// <para>示例值：1610281603</para>
         /// </summary>
         [JsonPropertyName("create_time")]
-        public int? CreateTime { get; set; }
+        public long? CreateTime { get; set; }
 
         /// <summary>
         /// <para>更新时间</para>
