+++ src/Ccm/PostDriveV1FilesByFileTokenCommentsBatchQueryResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-12
 // ************************************************************************
 // <copyright file="PostDriveV1FilesByFileTokenCommentsBatchQueryResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -55,7 +55,7 @@ public record PostDriveV1FilesByFileTokenCommentsBatchQueryResponseDto
         /// <para>示例值：1610281603</para>
         /// </summary>
         [JsonPropertyName("create_time")]
-        public int? CreateTime { get; set; }
+        public long? CreateTime { get; set; }
 
         /// <summary>
         /// <para>更新时间</para>
@@ -269,7 +269,7 @@ public record PostDriveV1FilesByFileTokenCommentsBatchQueryResponseDto
                 /// <para>示例值：1610281603</para>
                 /// </summary>
                 [JsonPropertyName("create_time")]
-                public int? CreateTime { get; set; }
+                public long? CreateTime { get; set; }
 
                 /// <summary>
                 /// <para>更新时间</para>
