+++ src/Helpdesk/GetHelpdeskV1FaqsByIdResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-12
 // ************************************************************************
 // <copyright file="GetHelpdeskV1FaqsByIdResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -108,7 +108,7 @@ public record GetHelpdeskV1FaqsByIdResponseDto
         /// <para>示例值：1596379008</para>
         /// </summary>
         [JsonPropertyName("create_time")]
-        public int? CreateTime { get; set; }
+        public long? CreateTime { get; set; }
 
         /// <summary>
         /// <para>修改时间</para>
@@ -192,7 +192,7 @@ public record GetHelpdeskV1FaqsByIdResponseDto
         /// <para>示例值：1596379008</para>
         /// </summary>
         [JsonPropertyName("expire_time")]
-        public int? ExpireTime { get; set; }
+        public long? ExpireTime { get; set; }
 
         /// <summary>
         /// <para>更新用户</para>
