+++ src/Helpdesk/GetHelpdeskV1FaqsResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-12
 // ************************************************************************
 // <copyright file="GetHelpdeskV1FaqsResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -140,7 +140,7 @@ public record GetHelpdeskV1FaqsResponseDto
         /// <para>示例值：1596379008</para>
         /// </summary>
         [JsonPropertyName("create_time")]
-        public int? CreateTime { get; set; }
+        public long? CreateTime { get; set; }
 
         /// <summary>
         /// <para>修改时间</para>
@@ -224,7 +224,7 @@ public record GetHelpdeskV1FaqsResponseDto
         /// <para>示例值：1596379008</para>
         /// </summary>
         [JsonPropertyName("expire_time")]
-        public int? ExpireTime { get; set; }
+        public long? ExpireTime { get; set; }
 
         /// <summary>
         /// <para>更新用户</para>
