+++ src/Hire/DeleteHireV1ExternalApplicationsByExternalApplicationIdResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-12
 // ************************************************************************
 // <copyright file="DeleteHireV1ExternalApplicationsByExternalApplicationIdResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -98,7 +98,7 @@ public record DeleteHireV1ExternalApplicationsByExternalApplicationIdResponseDto
         /// <para>必填：否</para>
         /// <para>示例值：1</para>
         /// <para>可选值：<list type="bullet">
-        /// <item>1：HR寻访</item>
+        /// <item>1：HR 寻访</item>
         /// <item>2：候选人主动投递</item>
         /// <item>3：人才推荐</item>
         /// <item>4：其他</item>
@@ -113,7 +113,7 @@ public record DeleteHireV1ExternalApplicationsByExternalApplicationIdResponseDto
         /// <para>示例值：1618500278645</para>
         /// </summary>
         [JsonPropertyName("modify_time")]
-        public int? ModifyTime { get; set; }
+        public long? ModifyTime { get; set; }
 
         /// <summary>
         /// <para>终止类型</para>
