+++ src/Hire/GetHireV1AttachmentsByAttachmentIdResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-12
 // ************************************************************************
 // <copyright file="GetHireV1AttachmentsByAttachmentIdResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -71,6 +71,6 @@ public record GetHireV1AttachmentsByAttachmentIdResponseDto
         /// <para>示例值：1618899376480</para>
         /// </summary>
         [JsonPropertyName("create_time")]
-        public int? CreateTime { get; set; }
+        public long? CreateTime { get; set; }
     }
 }
