+++ src/Hire/GetHireV1JobsResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-12
 // ************************************************************************
 // <copyright file="GetHireV1JobsResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -417,7 +417,7 @@ public record GetHireV1JobsResponseDto
         /// <para>示例值：1617170925462</para>
         /// </summary>
         [JsonPropertyName("create_time")]
-        public int? CreateTime { get; set; }
+        public long? CreateTime { get; set; }
 
         /// <summary>
         /// <para>更新时间 请使用update_timestamp</para>
