+++ src/Hire/GetHireV1ReferralsGetByApplicationResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-12
 // ************************************************************************
 // <copyright file="GetHireV1ReferralsGetByApplicationResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -55,7 +55,7 @@ public record GetHireV1ReferralsGetByApplicationResponseDto
         /// <para>示例值：1618899376474</para>
         /// </summary>
         [JsonPropertyName("create_time")]
-        public int CreateTime { get; set; }
+        public long CreateTime { get; set; }
 
         /// <summary>
         /// <para>内推人的 ID</para>
